postProcess delegates to item-type processors
