+++ packages/common/src/sharing/index.ts
@@ -1,0 +1,1 @@
+export * from "./share-item-to-groups";
