+++ packages/simple-types/package.json
@@ -13,8 +13,12 @@
     "dist/**"
   ],
   "devDependencies": {
+    "@esri/hub-common": "^4.0.0",
     "rollup": "^1.22.0",
     "typescript": "^3.9.2"
+  },
+  "peerDependencies": {
+    "@esri/hub-common": "^4.0.0"
   },
   "dependencies": {
     "@esri/solution-common": "^0.9.2",
