+++ packages/simple-types/src/simpleTypeHelpers/generate-empty-creation-response.ts
@@ -1,0 +1,11 @@
+import { ICreateItemFromTemplateResponse } from "@esri/solution-common";
+
+export function generateEmptyCreationResponse(
+  templateType: string
+): ICreateItemFromTemplateResponse {
+  return {
+    id: "",
+    type: templateType,
+    postProcess: false
+  };
+}
